requirements.txt
Flask==2.3.3
duckdb==0.8.1
pandas==2.1.1
requests==2.31.0
pyarrow==13.0.0
python-dotenv==1.0.0
openai==1.12.0
asyncio
